[FE] REFACTOR: mapInfo 코드 리팩토링 - slider이름 다시 원래대로 변경..
